requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
dnspython>=2.6.1
aiohttp>=3.9.0
tqdm>=4.66.0
